requirements.txt
pyyaml
dnspython
tldextract
openai
python-dotenv
python-whois
